literals

+++ src/main/c/frontend/bison/bison_actions.c
@@ -1,3 +1,3 @@
 #include "bison_actions.h"
 
-i32 parse_integer(i32 i) { return i; }
+i64 parse_integer(i64 i) { return i; }
